refactor: change submodule name to store
