+++ docarray/store/exceptions.py
