+++ docarray/store/helpers.py
