+++ tests/integrations/store/docker-compose.yml
